timeutil: fracsec size;

+++ libdariadb/timeutil.h
@@ -13,7 +13,7 @@ struct DateTime {
   uint8_t hour;
   uint8_t minute;
   uint8_t second;
-  int fracsec;
+  int64_t fracsec;
 };
 
 /// current timestamp with nanosecond.
